[proto] Port sdk documentation to asciidoc
Constraints were:
- HTML output
- PDF output
- Easy to edit
- Work on Windows

Remaining contenders were mostly asciidoc and sphinx. Sphinx seems to
require the dblatex toolchain to generate PDF which is a showstopper
on Windows. Asciidoc can do FOP.

Asciidoc is now packaged, see the README.txt included in the package
for details. The sdk.dox documentation was ported to asciidoc, the
following issues remain:
- Cross references were not fixed. Better to wait for the document to be reorganized first.
- mscgen and PDF do not interact well, generated images/graphs are too large. This should be fixed with mscgen 0.21, unfortunately binaries are not available and I could not build it. The issue was reported here:

  http://code.google.com/p/mscgen/issues/detail?id=75&thanks=75&ts=1346152959

I will keep the doxygen version until the issue is resolved.