[proto][doc] force mscgen charts width to fix PDF output
With the previous hscale="1.5" setting, the charts generated by mscgen
were not handled correctly by the PDF toolchain, the images were too
large for the selected display width and shift to right, ending
half-cropped in the PDF document.

The real problem is asciidoc/FOP/asciidoc-mscgen are not configured
correctly for the output document size. Embedding one of the generated
PNG or SVG into a test asciidoc with "autowidth" set gives a correct
PDF. So asciidoc-mscgen is probably the culprit here but I have not
found a way to force the right options in the filter yet.

For now, just force the output to 550px wide, there is a bit of
aliasing and the figures are smaller but it produces a usable output.