[mission_tester] remove project and sources
The project was shipped along with scipio but nobody uses it for
sword/trunk. QA has mostly rewritten a better version of it in C# and
is no longer maintaining this one. The code is said to be useless
complicated and buggy. Let's kill this thing and write a better one if
necessary.