[build] use generate-database/symbols as regular targets, remove antcall
A new configure-update target was created to execute all the package
updates. Then the generate-database-* were marked as depending on
configure-update and finally configure depends on configure-update and
the generate-database-*. This should achieve the same result than the
previous setup without requiring antcall.

Removing antcall makes it easier to analyze target dependencies.