[build] get QT translation files from QT package, not svn
The former behaviour was:
- The svn checkout contains qt translation files in
  data/app-data/resources/locales.
- The QT package contains qt translation files deployed in run/ by a
  dependency of generate-database.

The QT files are more recent and more importantly, are synchronized
with the QT release, but were missing the arabic translation
file. Before applying this change, the arabic file was generated from
QT sources with lrelease (it was matching the svn one).

This change removes the svn version and replaces them with the ones
brought by QT, using a new target ran by configure.