[build] concatenate all license files into license.dat
Since the applications are not being installed on the build servers,
Flex has no way to guess the license files should be picked from
the run directory, and even if it happens it is apparently after
resolving others from the Registry. "license.dat" is the default
license path defined in license_gui & co, so try our chance with
this one first.