[simulation_kernel] merge E_EngineReport and E_DecisionalReport
MIL_Report loads available reports from reports.xml. When the
simulation wanted to post a report, it was passing an E_EngineReport
value which was mapped to an E_DecisionalReport value which was
checked against loaded reports. The E_EngineReport to
E_DecisionalReport mapping was maintained manually. And had no real
value. Just get rid of it and maintain a subset of E_DecisionalReport
values.