[MT_Tools] specialize MT_InterpolatedFunction, remove templating
There is clearly no need for two TypeIn and TypeOut. And TypeIn has
always been double so let's accept we do not need this specialization
and use double directly. I have not moved the implementation in an
implementation file as this code is apparently used in tight loops by
the simulation and I do not have the time to profile again without
inlining.