[preparation_app] add --debug-dir option
When the option is passed, the directory is created and a file logger
bound to debugdir/preparation.log is registered. I am not fond of this
system, I would rather have an option to state explicitely where I
want the log file but simulation_app and other applications already
work that way.