[build] retrieve brain ide from a new masalife/ide package
Previously, the IDE was a zip in the masalife package. Unfortunately,
we have no idea how this zip was generated nor how we can reproduce
it. Still, we know how build masalife and are likely to delete and
recreate distribution directories in the course of development. But if
we do so, we risk losing the IDE zip. Therefore it feels saner to
separate both artifacts and put the IDE in its own place.