[simulation_kernel] SWBUG-9550: deduplicate tactical lines, again
When giving missions with tactical lines arguments, the tactical lines
points are copied in each mission protobuf message. On the simulation
side, we have to deduplicate them again based on their geometry before
registering them in the pathfinder threads. r36912 broke this
mechanisme by mistake, which this change reintroduce in a different
form.

In pratice, it meant the tick 10 of Benchmark_BattalionOpenTerrain
requested around 3000 duplicate lines crossing the whole east Paris
area to be registered, instead of 55.