[simulation_terrain] fix TER_LimitData::Distance
The algorithm looked like:

  for each segment of the line:
    if the projection of p on the segment is:
      - before the first point, return the distance to the first point
      - on the segment, return the height to the segment
      - after the second point, return max float

This is obviously wrong for any p which project is past every segment
last point. The fix is to change the last condition so if we are
evaluating the last segment of the line, the distance to the second
point is returned.

Merged revision(s) 37436 from branches/5.1