[simulation_terrain] break cyclic dependency with simulation_kernel
The culprit was some cleanup code which had to be run on
TER_PathFindRequest_ABC by the pathfinder, but required to call some
registration method on DEC_Pathfind_Manager. The registration logic is
related to the threaded implementation of the pathfinder and has
nothing to do in the DEC_Path_ABC hierarchy. The solution is to
introduce a real request class, separated from DEC_Path_ABC. When the
pathfind computation is done, the request cleanup method is called and
trigger the deferred cleanup code, which in turn will call the actual
cleanup code on DEC_Path_ABC.

In practice, it means DEC_Path_ABC no longer inherits from
TER_PathFindRequest_ABC, and DEC_Pathfind_Manager registers custom
requests inheriting TER_PathFindRequest_ABC, holding a DEC_Path_ABC
and implementing the deferring logic.

With this, simulation_terrain no longer depends on simulation_kernel.