[preparation_app]Fix Jira SWBUG-9566 Plantage prépa sur drag&drop unit sur un fantôme. Resolved qt selection problem when replacing a ghost by a unit on the map OR in the orbat.
Problem source:
Error was due to a "select" event sent after the use of the qt function takeRow( int row ) when the row in parameter was selected before being taken.

Solution:
Set selection on the new unit replacing the ghost before the deletion of the ghost on the tacticaltreeview.
Merged revision(s) 37522 from branches/5.1: