[adn] Jira SWORD-1462 [Part IV]: Symbols are part of the physical database
Symbols directory in the physical database is used to load symbols.
If this directory is not found, ADN extract the content of symbols.pak to the physical database directory.

A xslt transforms old physical.xml to add the symbols-directory entry.

A consistency warning is displayed if some symbols are defined in symbols.xml and are not present in the symbols directory.

Improve symbols management in ADN, GLSymbols is now defined only one time.