[simulation] SWBUG-9571 changed simulation test mode option to step ticks
It's hard-coded for now because the whole session.xml/command line option system
should be refactored to make it possible to override session parameters from the
command line.

The code is much more similar now between a checkpoint and a regular start.
Plus it fixed a synchronisation issue where the simulation was in 'wait' state
after loading a checkpoint and couldn't be made to run.

A side effect is that now the paused attribute in the session.xml is used when
loading a checkpoint whereas before it would imply the simulation was to be
reloaded in a 'paused' state.