[clients] Fix Jira SWBUG-9774: Crash ODB logistique
Overload contextMenuEvent in gui::LogisticTreeView to handle clic on logistic level.
This TreeView is the only one that inherit EntityTreeView and doesn't stock only Entity_ABC elements, but someday we could have more.

Todo: Improve StandardModel by checking if IsA< T > in the beggining of GetDataFromItem.
This check invalidate all our actual treeview, because Agent_ABC are registered as Agent_ABC and retrieved as Entity_ABC.
To fix that, we will need a way to register each entity with several mimeType.
We should:
  - add a RegisterToStandardModel method into GraphicalEntity_ABC (best way, not possible now because of crossed reference between clients_gui & clients_kernel)
  - specialize template method in StandardModel to make each needed registration for each possible type (may be problematic with high level classes (HierarchyTemplate in preparation.lib for exemple)
  - use a dynamic_cast to check if Entity_ABC (not very efficient if one day we need more sub type, or another second type than Entity_ABC)
  - some other way ?