[gaming][simulation] Jira Story SWORD-1473: Phase change reports can be translated
In order to translate stage changes, we differentiate reports parameters of type stages from those of type string.
We also add a Stages.xml to each Physical bases to store the effective translation.
ADN keep Stages.xml model-version up to date.

So, when gaming receive a report with a parameter of type stage, it check the Stage.xml file to find a translation.
If the translation exist, we display it, elsewise we display the original token (un-translated).