[dispatcher_dll] log dispatcher configuration errors in dispatcher.log
This is the usual chicken and egg problem, the log file location is
retrieved from some SessionConfig object which may fail to initialize
when the XML configuration is incorrect. The fix is not really one but
will be tested in gosword: we assume that before reaching the XML
parsing routine all details required to create a Dispatcher.log are
available. This is at least true for its location (not for the log
file size limits and similar options), so the error is appended to a
possibly existing log file before termininating the dispatcher.