[process] make SimProcess.Wait() more reliable
It is possible that os.Process.Wait() is thread-safe though it is not
clear from the documentation, but SimProcess requires cleanup upon
exit anyway. So wrap the system Wait() in a goroutine waiting for
termination and handle wait requests with channels.

As a side effect, when Wait() times out it no longer terminates the
process.