[process_test] avoid a race in TestSuccess()
StartSim returns success if the simulation starts and we manage to
connect to it within a timeout. But between these events, the
simulation can run fast enough to finish its ticks and exit
succesfully. Until we have a way to unpause it there is no reliable
way to check anything but the exit status.