[model] add a Model structure
The model registers a handler to a Client and updates itself with
incoming messages in a separate goroutine. It supports two kind of
actions: commands which are just functions applied once on the Model,
mostly used to implement queries, and conditions which are fired each
time a message is processed until they return true.

Only IsReady() and WaitReady() methods are implemented at this point, to
wait for state initialization to finish.