[client] close current handlers upon disconnection
When remote drops the connection, listen() has to notify serve() to
abort currently registered handlers and to immediately invalidate ones
being registered. Otherwise the caller might wait on a command until it
times out.

Also, move back handlers cleanup in serve() loop as the goroutine is
responsible for the map, not Close(). This is important now the cleanup
could be triggered from different threads.