[package_app] remove unused application, call selftraining_app directly
As far as I know, package_app is only used to bind .otpak files to it,
which in turn calls selftraining_app --install package. Let's register
the latter directly as .otpak handler.

For this to work correctly, selftraining_app.exe must be spawned with
its working directory set to the installed "applications" directory,
or it fails to find wome resources. To achieve this in a restricted
call context (Windows registry), a --cwd option is added to tell
selftraining_app.exe to change it before executing anything else.