[simulation_kernel] add custom result field to UnitMagicActionAck
UnitMagicAction has custom parameters which means it cannot have a
generic reply. Instead, add an output MissionParameters result
field. The first action to use it is FormationCreation which sets the
created formation identifier. It means callers no longer have to watch
for a tagged FormationCreation message to know which formation they
created.