[model] serialize message, command and condition requests
This might increase a bit the latency of command and condition requests
but it gives us some analogy of memory barriers for messages: if we
issue a command requests from a MessageHandler, we will be sure that all
messages preceding the one being handled will have been processed by the
model. This is a very interesting guarantee when writing Client
commands.