[dispatcher] add and handle DisconnectionRequest message
When running tests connecting and disconnecting clients very quickly it
may happen that the dispatcher does not pick a disconnection and
decrement the connected client count before a new client is trying to
connect to it, ending in a connection rejection. This message ensures
the client disconnection is taken in account before moving on to the
next test.