[dispatcher] make ClientsNetworke reports only the first ConnectionError
We assume that ConnectionError() is only fired after a
ConnectionSucceeded() or ConnectionFailed() event was triggered. This
ensures the connection is registered and the ConnectionError() message
will be displayed once and only once.

Before this change, disconnecting a client server side could result in
many ConnectionError() being logged, one for the disconnection
exception, one for the disconnection itself and others for aborted
writes.