[tools] make ExtractArchive extract empty files, fix nonsensical API
For some reasons the zipstream wrapper does not like empty buffer, and
initialized itself at !good() on them. Just avoid it and use zipfile
directly.

Also, the "pass NULL filename" to extract the current file API was a
little crazy, so I rewrote the zip extraction with an enumerator,
which is how izipfile works under the hood anyway.