[model_test] make printer rewriter unstable properties
Entity identifiers may change when the physical database change. For
instance, report templates identifiers belong to the global identifier
namespace which means each time a report is modified, our identifier may
shift. To avoid this, add a prettyPrinter mode where identifiers are
replaced with '-'. The regular mode is still useful for debugging
purpose.