[build] update module version info with a custom_target
The goal is to avoid having to relink all modules every time the version
number changes, and update the binaries directly which is much faster.
To achieve this, a binary resource templates is loaded and its
RT_VERSION component copied in target binary. Then we update it with the
desired values.

editres.py was update with new commands:
- print prints the version resource in a python text format, useful for
  debugging.
- dump serializes target module resources in python pickle format
- copy copies selected resource types from a source which can be another
  module or a pickled input, into target binary