[client] replace ClientId with GetClientId(), add commands
Client.clientId is managed by serve() goroutine. Instead of trying to be
smart and manage two attributes one read and written by the goroutine
and the other read and written by the client, it is much safer and
easier to keep the serve() one and query it via the usual mix of
closures and channels.