[simulation_terrain] replace TER_DynamicData logic with shared_ptr
TER_DynamicData is a combination of a (points, terraindata) tuple and a
referenced counted object tracking which thread is referencing it. It
costs one mutex and one map per dynamic data. Instead, a simpler
solution is to let the threads track the dynamic data themselves. Also,
it isolates the retractation handler business deeper into
TER_PathFinderThread.