[login_test] make TestNoDataSentUntilSuccessfulLogin more useful
This was one of the first test written and at that time it could only
check that a connected but not authenticated user was not receiving
initialization/heartbeat messages. Now we can connect it while
performing actions with other users. The first change to this test is to
keep listening while shutting down the simulation, which used to fail
since the simulation was sending profile destruction messages.