[errors] introduce SwordError and a related gocheck Checker
This change introduces a new swapi.SwordError error struct representing
a protobuf error response. It contains an integer code, an optional
error message and a convenience Name string generated from the error
code. Everything else sums up to using this class and adjusting tests:

- Rename nameof() into makeError(), and extend it to return SwordError
  when possible.
- Add a gocheck checker to check an argument is a SwordError and its
  Name is a supplied string.

I would have liked makeError() to be simpler but the use of enumerations
for protobuf error codes prevents me for using an interface to get the
error code from the response message.