[sim] change_knowledge_group party identifier is now deprecated
The current constraint is an automat must belong to the same party than
its knowledge group, so we can use the automat party to lookup the
knowledge group. Anyway, relaxing this constraint would be cheap, either
loop on the parties (few parties, cheap lookups) or merge knowledge
groups in a single container.