[crowd_test] make teleport test more reliable
The problem with the simulation is it exposes a lot of "invalid" state
through the protobuf. For instance, when teleporting a crowd, it first
creates a new empty concentration, then empty the existing flow and
concentration, then fill the new concentration, then destroy the former
flow and concentration. So testing that we have 1 flow or concentration
is not enough, it might depend on the order of those operations.