[tools] Remove virtual destructor from Message.h
Because there are no subclasses so it's useless and wastes space, and because sometimes the application crashes and ends up in this virtual destructor (probably because the memory has been scratched somewhere else) (SWBUG-10598 for instance). It won't fix the crash, but maybe we'll get stack traces which make more sense if the corrupted memory is not in this class that gets created very very often.