[simulation] SWORD-1894 vision cones are now sent right away by the simulation when the dispatcher registers to receive them
Previously the simulation needed a tick and this would incur a delay for the first client to register because the subsequent clients receive the initial vision cones from the dispatcher cache anyway.
Also the new code removes the need to query whether the vision cones must be sent for initialization every tick.