SWORD-1912 reworked UI for configuring logs
We now have a check box to toggle log rotation on and off.
Also the maximum size of a log file cannot be 0 anymore (this is used to deactivate the logs all together, but we don't want to do it, and if we did it would probably be with a check box).
And finally the maximum number of files now sets the number of rotated files to keep excluding the main log file whereas before it was including the main log file.
Only the current log files are displayed as tabs in the log page of a session.