[simulation][gaming] Fix JIRA # SWBUG-10756 Fix posture algorithm
The posture display in gaming does not update itself anymore and does not fill the last segment when a unit is parked in an engineer prepared area
The problem was that posture computed by the simulation ended at poste_->posteAmenage_/0% progress, which was the same as poste_, so posteAmenage_ effects were never applied.
The user interface worked around this by adding a +1 to compensate, but it didn't show correct result for postePrepareGenie_ because this stance was completed at 100% (so it worked, but the work-around in gaming broke its display).
I changed the code so all the logic is done in DefaultPostureComputer and the role just retrieves the result.
The test is changed accordingly to emulate the way the role uses the computer.