[adn] fix crash on close
To explain that crash:
ADN_MainTabWidget is the parent of each ADN_Gui_ABC's mainWidget_.
These mainWindgets are created by CreateScrollArea, for the most part it's a QSplitter (now an ADN_SmartSplitter).
That QSplitter is the parent of the SearchListView and of the scrollArea (the right part of each tab).

When ADN_MainWindow reset the ADN_MainTabWidget pointer, it delete all his chidren, so each QSplitters.
When a QSplitter is deleted, his children are deleted in the opposite order of their insertion.
To have the list view on the left and the scroll area on the right, we need to add them in this order.
So the scroll area was deleted before the list view (ADN_SmartSplitter fix that).

When deleting the list view, it start by purging it's content, and that emit the signal selectionChanged.
In some tabs (units and equipment at least), we connect the selectionChanged signal of it's list view to a slot owned by the ADN_XXX_GUI.
When the signal was emitted by the deletion of the list view, the ADN_XXX_Gui is still alive (it's the ADN_Workspace::Purge() method that delete it, and it's call after the deletion of the ADN_MainTabWidget),
but in this slot, the ADN_XXX_GUIs attempt to make some change on elements that were owned by the scrollarea, so dead elements if the scrollarea was already destroyed.

ADN_SmartSplitter delete the list view before other children, so the scroll area is still alive when the list view emit it's last selectionChanged signal.

It should have been fixed by adding a Disconnect method to ADN_GUI_ABC that we would have call before reseting the ADN_MainWindow, and overload it in each tab to do the required disconnect.
Doing this would have been much longer, bad for code readability by adding a method to each ADN_XXX_GUI, with potential forgotten signals...
I think this fix is simplier and more efficient.