[gaming] fix profiling panel out bytes count
Output bytes are counted after they are queued not after they are
actually sent, this is good enough for a profiler. I am not really fond
of storing the counters in Socket but it is already done that way for
the number of sent messages. It will not be too hard to change.