[adn] fix logistic tab display:
This is an singularity: on others tab, the data are connected when we select an element from the list view.
Here (on the logistic tab), without a list view, the data are initialized and connected before the gui is built.
Calls to ADN_Connector_Vector::AddItem are done without a gui element, so it can't refresh a gui element that doesn't exist.
We need a final AddItem( 0 ) when building the gui to force the refresh of these table views.

The big refactor on adn's load process introduced that bug.