[tools] separate DefaultLoader and PhyLoader
I was a little optimistic when I made ExerciseConfig::GetLoader
unavailable before the exercise is loaded. Loader_ABC was used for 2
different things: loading random XMLs and loading physical files. The
latter are only available once the decisional is known which happens
once the exercise is loaded. The former only requires the application
run directory to be defined. A lot of code was abusing ExerciseConfig to
load random XML before the exercise was initialized.

The solution is to clearly separate both roles. PhyLoader no longer
inherits DefaultLoader and Loader_ABC but embed them. As a bonus,
Loader_ABC unimplemented methods can be removed from DefaultLoader.