[adn] stop removing single quotes from image directory names
Xalan xsl::param API is weird borderline buggy. If you put unbalanced
single quotes in values it whines about it as if it had to be surrounded
like select="..." expressions. If you surround the whole expression it
fails differently. If you escape them, XSL attribute templates escape
them again. The only working technique I have found is to build the HTML
tags with xsl::text and make use of disable-output-escaping.