[gaming] do not toggle vision cones until connected
Sending vision cones toggling message before being connected fails with
error_invalid_parameter which should really be error_forbidden. This is
harmless but writes 3 errors lines in gaming log at every connection.
Here we listen to Profile and Simulation events and check both are
connected before allowing cones toggling. It would have been nicer to
listen only to Profile but it breaks a lot of other things in gaming if
a NotifyUpdated is added in Profile::Clean().