[clients_kernel] override readLocale() with MASA_LOCALE instead of readLang
Use it like:

  MASA_LOCALE=en_US adaptation_app.exe [...]

It is more powerful than the previous version since it drives the locale
rather than the language, but will not override locale passed
explicitely to AddTranslator. Currently this is only done by the
logistic_plugin, we could find a better solution or force the override
when we need to test it.