[protocol] extract magic actions mappings from clients_kernel
I have chosen the primitive types version because its behaviour is well
defined, and because it supports static asserts at least within a
module. I am not completely sure this is the good solution though.
Editing it may be a bit tedious and the definitions lack locality.
Anyway this is already better than having a giant if/then/else.

All the enumeration business left in MagicActionType comes from us
saving a translated version of the enum values (sometimes) and having at
least 4 different ways to turn enum values into strings.