[sim] add send_state option to checkpoint message
This makes the simulation send its full state right after the checkpoint
prologue message. This is a test feature, used by callers to get the
simulation state actually saved in the checkpoint, so it can be compared
to the one restored from the checkpoint.

If nothing is done, many plugins including the dispatcher model reset
with all kind of unwanted side-effects. To avoid this, a new plugin is
added at the root of the plugin hierarchy to detect the stream of
simulation state messages, prevent it from reaching other plugins while
forwarding it to source client.