[sim] make sure a reloaded, paused simulation does not tick
This property is a must to automate checkpoint tests. Combined with
checkpoint send_state option, it lets us get the state before the
checkpoint and compare it to the state reloaded from the checkpoint,
from a protobuf point of view.

Before this change, a paused simulation would tick or half-tick once
when reloaded from a checkpoint.