[dispatcher] turn the plugin list into a tree, refactor filtering
The original motivation is to solve the following problem:
- Vision cones messages must not be broadcast by ClientsNetworker but are
  to be sent to specific clients.
- A (yet-to-be) simulation state message sequence must be sent to a
  single client and masked to all others as well as plugins.

Vision cones implementation was specific to ClientsNetworker and a bit
invasive to configure.

The initial approach was to add a return value to Receive(*) methods to
signal the message should not be forward further. But conflicting
constraints arise very quickly in this kind of scheme.

The current approach is to extend PluginContainer a bit and turn the
plugins chain into a tree. That way, the vision cones plugin becomes a
parent of ClientsNetworker and can prevent messages to reach it and only
it. The simulation state problem will be solved by adding a root plugin
filtering the state for all others.

Conflicts can still appear in complicated situation but are easier to
solve on case by case. Finally, the plugin tree is defined explicitely
in PluginContainer which gives a clearer view of what is happening.