[build] make dispatcher_dll depend on plugin DLLs
The motivation is to force plugins to be rebuilt when compiling server
targets like simulation_app. Without this, simulation_app may build
against a dispatcher version, and a Plugin_ABC version, not matching the
plugin DLLs, causing all kind of crashes and weirdness.

The trick is to make dispatcher_dll depend on the plugin DLLs. Since
server application depends on dispatcher_dll, it simplifies the
CMakeLists.txt.