[gaming] Fix JIRA # SWBUG-11252 Gaming - Limits management is not clear
If a limit is selected, filtering behaves as if it was a subordinate of the associated entity.
'Selected' shows limits defined by superiors of the selected entity or the selected entity.
'Superior selected' shows limits defined by superiors of the selected entity, the selected entity, or immediate subordinates of the entity.
'Controlled' shows limits visible by the profile (as limits can be created on these).
'Always' and 'Never' unchanged.