[gaming_app] tweaked large roads, motorways and bridges for mapnik
Different roads seem better connected that way and the fact that their widths are less different means the bridges don't look too ugly.
Still it's difficult to have a clean integration for bridges (even more than for tunnels) because there is no way to make the difference between a railroad bridge and a motorway bridge.