[logistic_plugin] add history querying api
This commit has two parts:

1- Introduce a new way for plugins to handle client messages.
Previously, plugins were registering themselves to ClientsNetworker and
had to know an AuthenticatedLinkResolver_ABC if they wanted to answer to
a single client. The new version adds a HandleClientToSim() method to
Plugin_ABC, receiving the message and two publishing interfaces, one for
unicasting the other for broadcasting. The advantages are:
  - Use standard interfaces, plugins do not have to know about
    ClientsNetworker.
  - Hide AuthenticatedLinkResolver_ABC
  - Enable input message filtering through PluginContainer hierarchy

What is not supported is plugin emitting messages asynchronously. This
could be done by registering a regular ClientPublisher_ABC at plugin
instanciation, again hiding ClientsNetworker.

The RewritingPublisher_ABC is an unfortunate side-effect of
ClientPublisher_ABC taking const messages. It lets us handle the
client_id/context business out of the plugins. This interface could be
generalized at handler level, and ClientPublisher_ABC be used only at
networking level, or removed completely.

2- Add a minimalistic API to fetch logistic history. Logistic request
identifiers are passed and related history strips returned,
deduplicated. Pagination options can be added later, the server will
limit the response by itself.

Note the error handling is done in parent SimulationToClient message.
The motivation is I have to add another message to control the logistic
history size. But the new message will have to define another way to
return errors, probably with an Ack message and an error code. Instead,
I would rather *not* create a new Ack message and just return an error
in the parent message. Other messages errors will be slowly migrated to
this scheme. Error codes may be introduced if we find them useful.