[gaming] disable history toggling checkbox until its fate is decided
The filtering feature seems implemented in gaming but it is not really
useful since consigns are deleted in the logistic plugin and gaming upon
destruction message. But it still does not work if the deletions are
commented. I have not enough time to look closely now, I am keeping the
code for later.