[gaming_app] added road names to mapnik layer
Right now there's a test to filter out road names >= 5 characters as the technique used works for names such as A104 or D123B and isn't that great for longer names.
The main reason is that it uses a halo around the letters to simulate a shield because the usual open street map shields are tied to a given color hard-coded in their mapnik configuration file.
As we have changed the road colors we cannot use out of the box those shields, which anyway typically come in only 8 widths (ranging from 1 character to 7 characters long) and as such only work for names shorter than 9 characters.