[logistic_plugin] build and archive LogHistoryEntry messages
We want to record handling creation/update/destruction messages so they
can be indexed and retrieved by history requests. Each ConsignData_ABC
now have a LogHistoryEntry which is updated with new messages and
immediately serialized in a new binary archive file, rotating on file
size.

LogHistoryEntry represents the *full* state of a request at a given
point in time. Possible fields combinations are:
- creation -> newly created handling
- creation + update -> regular handling
- creation + update + destruction -> handling state before destruction

This representation is more space consuming than just recording the
creation/update/destruction but it lets us truncate the history and
avoids long seeks to retrieve the initial creation message.

Reusing existing messages in LogHistoryEntry reduces the amount of
serialization code to write and helps with maintenance.

The tests were modified:
- To ignore the new archive file
- To make input message valid and serializable