[gaming_app] split road casing style into several style, one for each road type
It doesn't change performance but introduces some flexibility in the ordering of styles applied.
Also in a near future we will likely generated different shapefiles for different road types to increase performance as for instance  it won't be necessary to iterate over a potential huge road set just to draw the motorways.