[dispatcher] avoid "forbidden" ack on logistic_plugin messages
The logistic_history_request message is handled by the logistic plugin
but was unknown to the DispatcherPlugin/Profile, which sent a
"forbidden" ack upon reception. The quick fix for IITSEC is to allow the
message at Profile level.

A future better fix is to make DispatcherPlugin use HandleClientToSim
instead of registering OnReceive. It would not solve the case when the
DispatcherPlugin is not loaded but this is a real issue anyway.