[logistic_plugin] reimplement history requests with LogHistoryEntry
The previous version was maintaining the history of live requests in a
structure mapping the common denominator of all requests. This had two
problems:
- It was not a real history, you could not fetch the history of finished
  requests.
- The representation was restrictive.

The new version plugs the LogHistoryEntry archives to the requests
history protobuf command. This fixes the issues described above at the
price of a more complicated representation. But reusing existing
messages makes the serialization/deserialization easier at the same
time.