[gaming_app] filtered out "UNK" from road names
It's not defined in the shapefile format specification but it looks like a pattern in several data sources.
The portoshima terrain data are littered with "UNK" so this is at least an ad hoc solution here.
It's worth noted that the terrain project checks for "UNK", "unk" and "N_P" on different occasions.