[simulation] Fix JIRA # SWBUG-11361 Scipio - Foule bloqué par une barricade mais pas de CR ni d'ajout dans le groupe de connaissance foule.
Objects must hinder when moving from outside to inside but not from inside to outside.
The problem was that :
ponctualObjectsOnSet_ is a set of objects 'on the border' between steps but doesn't know whether the next object is inside or outside.
An additional step is added right at the border of objects, so the IsInside test was always true in that case.