[client] ensure that Client and Model goroutine terminates in Close()
This fix is in reaction to yet another termination bug in Client: when
the listening goroutine was terminating it posted on the unbuffered quit
channel. Unfortunately, the receiving serve() goroutine had often
terminated at this point and listen() was left hanging on some receiver.
It might explain the tests memory leak, which I do not observe locally.

Waiting for goroutines to terminate is not required but will help
spotting these bugs faster. I did not see any performance impact running
the test suite.