[sim] ensure segment subdivision on elevation grid is close to the original
The previous version was computing the elevation cells crossed by a
segment and was inserting intermediate points which were the bottom left
corners of cells. The path finder was then evaluating the slopes between
such points. We need something more refined as the evaluated slope was
sometimes really different than the one expected looking at segment
points.

The new version reuses the Bresenham implementation but this time over a
1m integer grid instead of the elevation grid. Intermediate points are
generated at each intersection of the segment and the grid lines
orthogonal to its major axis.

For instance, with cellSize == 100, the segment (50, 50) -> (250, 150)
will be split into:

 (50, 50) -> (100, 75)
 (100, 75) -> (200, 125)
 (200, 125) -> (250, 150)

Note that no split occurs on the grid line y == 100.

That way we ensure intermediate points lie nearby the original segment.
We do not compute "all" grid intersections because any cell tied to an
intersection we ignore is also tied to another intersection we
evaluate. Assuming elevations are obtained via bilinear interpolation or
something similar, these cells contributions will be take in account.